# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ontosyx API",
    "description": "The Semantic Orchestrator — Knowledge Graph Lifecycle Platform.\n\n## Response envelope\n\nEvery successful 2xx response is wrapped in:\n\n```json\n{ \"data\": <T>, \"pagination\": {\"next_cursor\": \"...\"}?, \"meta\": {...}? }\n```\n\nEach handler's `responses(... body = T)` documents the type of `data`, **not** the wire shape. Generated clients should either unwrap `data` in a post-processing step (the approach `web/src/lib/api/client.ts` takes — frontend callers receive `T` directly) or wrap every response type in the envelope shape when consuming the spec from a third-party codegen.\n\nError responses use a separate envelope: `{ \"error\": { \"type\": \"...\", \"message\": \"...\" } }` (see the `ErrorResponse` schema).",
    "license": {
      "name": "MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/projects/{id}": {
      "get": {
        "tags": [
          "Projects"
        ],
        "operationId": "get_project",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Project ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Project details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Project not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "$ref": "#/components/schemas/ErrorBody"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Projects"
        ],
        "operationId": "delete_project",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Project ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Project deleted"
          },
          "404": {
            "description": "Project not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "$ref": "#/components/schemas/ErrorBody"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/projects/{id}/apply-reconcile": {
      "post": {
        "tags": [
          "Projects"
        ],
        "operationId": "apply_reconcile",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Project ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectReconcileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Reconcile decisions applied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectRefineResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid decisions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "$ref": "#/components/schemas/ErrorBody"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Project not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "$ref": "#/components/schemas/ErrorBody"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Reconciled ontology invalid",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "$ref": "#/components/schemas/ErrorBody"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
    "/api/projects/{id}/design": {
      "post": {
        "tags": [
          "Projects"
        ],
        "operationId": "design_project",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Project ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectDesignRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Ontology designed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDesignResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or large schema gate",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "$ref": "#/components/schemas/ErrorBody"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Project not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "$ref": "#/components/schemas/ErrorBody"
                    }
                  }
                }
              }
            }
          },
          "504": {
            "description": "LLM timeout",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "$ref": "#/components/schemas/ErrorBody"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
    "/api/projects/{id}/refine": {
      "post": {
        "tags": [
          "Projects"
        ],
        "operationId": "refine_project",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Project ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectRefineRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Ontology refined",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectRefineResponse"
                }
              }
            }
          },
          "400": {
            "description": "No runtime or additional context",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "$ref": "#/components/schemas/ErrorBody"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Project not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "$ref": "#/components/schemas/ErrorBody"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Uncertain reconcile matches",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "$ref": "#/components/schemas/ErrorBody"
                    }
                  }
                }
              }
            }
          },
          "504": {
            "description": "LLM timeout",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "$ref": "#/components/schemas/ErrorBody"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "type": "object",
        "required": [
          "type",
          "message"
        ],
        "properties": {
          "details": {
            "description": "Optional additional details"
          },
          "message": {
            "type": "string",
            "description": "Human-readable error message"
          },
          "type": {
            "type": "string",
            "description": "Machine-readable error type (e.g., \"not_found\", \"bad_request\")"
          }
        }
      },
...
      "ProjectDesignResponse": {
        "type": "object",
        "required": [
          "project"
        ],
        "properties": {
          "project": {
            "type": "object"
          }
        }
      },
...
      "ProjectReconcileRequest": {
        "type": "object",
        "required": [
          "revision",
          "reconciled_ontology",
          "decisions",
          "uncertain_matches"
        ],
        "properties": {
          "decisions": {
            "type": "array",
            "items": {
              "type": "object"
            },
            "description": "User accept/reject decisions for uncertain matches."
          },
          "reconciled_ontology": {
            "type": "object",
            "description": "Reconciled ontology with user decisions applied."
          },
          "revision": {
            "type": "integer",
            "format": "int32"
          },
          "uncertain_matches": {
            "type": "array",
            "items": {
              "type": "object"
            },
            "description": "The uncertain matches being decided upon."
          }
        }
      },
...
      "ProjectRefineResponse": {
        "type": "object",
        "required": [
          "project",
          "profile_summary",
          "reconcile_report"
        ],
        "properties": {
          "profile_summary": {
            "type": "string",
            "description": "Summary of graph profiling results."
          },
          "project": {
            "type": "object"
          },
          "reconcile_report": {
            "type": "object",
            "description": "Report on ID reconciliation between original and refined ontology."
          }
        }
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    }
  },
...
}